Rename tool_parser files to remove _parser suffix
