+++ sgl-router/src/tool_parser/parsers/glm4_moe.rs
