+++ sgl-router/src/tool_parser/parsers/json.rs
