+++ sgl-router/src/tool_parser/parsers/kimik2.rs
