+++ sgl-router/src/tool_parser/parsers/mistral.rs
