+++ sgl-router/src/tool_parser/parsers/pythonic.rs
