+++ sgl-router/src/tool_parser/parsers/qwen.rs
