+++ sgl-router/src/tool_parser/parsers/step3.rs
